feat: add 'columns' utility and update docs
Added a new 'columns' utility to the index.ts file and created a corresponding 'columns.ts' file. Additionally, made updates in the README.md to illustrate the use of this new utility with examples. The 'columns' utility enriches the styling flexibility and these changes in the documentation makes it easy to understand its usage